fix(application): Fix remaining 11 compilation errors
- MakeChoiceUseCase: Use CompassChangeDto.Axis (not AxisId) and CompassChange.AxisId (not Axis)
- MakeChoiceUseCase: CompassDelta is non-nullable double, default to 0.0
- GetSessionStatsUseCase: Handle nullable StartTime in Subtract calls
- StartGameSessionCommandHandler: Use targetAgeGroup.Value for string assignment
- BadgeAwardingService: Add null check for axis in NormalizeAxisKey
- GetProfileBadgeProgressQueryHandler: Handle nullable axisId
- CheckAchievementsUseCase: Cast float threshold to int for ThresholdValue
- ExportCharacterMapUseCase: Handle nullable Metadata with empty dictionary fallback

+++ packages/application/Mystira.Application/UseCases/GameSessions/CheckAchievementsUseCase.cs
@@ -69,7 +69,7 @@ public class CheckAchievementsUseCase
                         IconName = !string.IsNullOrEmpty(axisBadge?.ImageId) ? axisBadge.ImageId : $"badge_{compassTracking.Axis}",
                         Type = AchievementType.CompassThreshold,
                         CompassAxis = compassTracking.Axis,
-                        ThresholdValue = threshold,
+                        ThresholdValue = (int)threshold,
                         EarnedAt = DateTime.UtcNow
                     });
                 }
